+++ zspotify/worker.py
@@ -1,8 +1,10 @@
 import sys
 import traceback
+import logging
 from PyQt5.QtCore import QObject, pyqtSignal, QRunnable
 from const import TRACKS, ARTISTS, ALBUMS, PLAYLISTS
 
+logger = logging.getLogger(__name__)
 
 class WorkerSignals(QObject):
     finished = pyqtSignal()
@@ -29,6 +31,7 @@ class Worker(QRunnable):
 
     def run(self):
         try:
+            logger.info("Attempting to start worker thread.")
             if "update" in self.kwargs.keys():
                 result = self.fn(
                     self.signals.update.emit, *self.args, **self.kwargs
@@ -40,6 +43,7 @@ class Worker(QRunnable):
         except:
             traceback.print_exc()
             exctype, value = sys.exc_info()[:2]
+            logger.error(f"{exctype} : {value} - {traceback.format_exc()}")
             self.signals.error.emit((exctype, value, traceback.format_exc()))
         else:
             self.signals.result.emit(result)  # Return the result of the processing